requirements.txt
Django>=4.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
drf-yasg
python-json-logger
openai
django-cors-headers

